+++ js/user-image.js
@@ -1,0 +1,19 @@
+const FILE_TYPES = ['jpg', 'jpeg', 'png'];
+
+const fileChooser = document.querySelector('.img-upload__input');
+const preview = document.querySelector('.img-upload__preview img');
+
+fileChooser.addEventListener('change', () => {
+  const file = fileChooser.files[0];
+
+  if (!file) {
+    return;
+  }
+
+  const fileName = file.name.toLowerCase();
+  const matches = FILE_TYPES.some((fileType) => fileName.endsWith(fileType));
+
+  if (matches) {
+    preview.src = URL.createObjectURL(file);
+  }
+});
